refactor: move react aria user event functions
